+++ tests/testthat/test-deprecated.R
@@ -4,5 +4,7 @@ test_that("deprecated functionalities generate informative errors", {
 
     aut$set_corresponding_authors(1)
     aut$get_contact_details(format = "{name} {details}")
+
+    orcid()
   })
 })