+++ NEWS.md
@@ -10,6 +10,8 @@
 
 ### Newly deprecated
 
+* `orcid()` is deprecated in favour of `icn_orcid()`.
+
 * The `format` parameter of `$get_contact_details()` is now deprecated in favour of the more explicit `template` parameter.
 
 # plume 0.2.5